add jsdoc stuff
