+++ index.html
@@ -1,12 +1,12 @@
 <!DOCTYPE html>
 <html>
-<head>
-  <meta charset="UTF-8">
-  <title>Monopoly</title>
- </head>
- <body>
-    <script src="bundle.js"></script>
- </body>
- <footer>
- </footer>
+  <head>
+    <meta charset="UTF-8">
+    <title>TableTop.JS</title>
+  </head>
+  <body>
+    <a href="/out/Game.html">
+  </body>
+  <footer>
+  </footer>
 </html>
