--- src/monopoly/cards/card.js
@@ -1,6 +1,0 @@
-function Card(text, action) {
-  this.text = text;
-  this.action = action;
-};
-
-module.exports = Card;
