+++ src/player.js
@@ -1,5 +1,5 @@
 /**
- * Represents a Player.
+ * A Player class
  * @constructor
  * @param {string} name - The player's name.
  * @param {int} number - The player's number.
@@ -12,6 +12,10 @@ function Player(name, number) {
   this.color = number;
 };
 
+/**
+ * Represents a Player.
+ * @param {int} position - The new position (index on the board) to move the player to
+*/
 Player.prototype.moveTo = function(position) {
   this.position = position;
 };
